V13: Move bucket assignment computation to reducer.cpp
Differential Revision: D14580911
Differential Version: 78604488
