+++ torch/csrc/distributed/c10d/ddp.cpp
@@ -1,7 +1,6 @@
 #include <torch/csrc/distributed/c10d/ddp.h>
 
 #include <torch/csrc/cuda/comm.h>
-#include <torch/csrc/utils/hash.h>
 #include <torch/csrc/utils/tensor_flatten.h>
 
 #ifdef USE_C10D_NCCL
@@ -38,27 +37,6 @@ void copyBroadcastTensorsToReplicas(
     }
   }
 }
-
-// Tensors may be coalesced into buckets. Buckets must contain tensors of
-// the same type, on the same device, so a bucket can identified by a
-// composite key of a tensor's type identifier and its device.
-struct BucketKey {
-  BucketKey(c10::ScalarType type, c10::Device device)
-      : type(std::move(type)), device(std::move(device)) {}
-
-  const c10::ScalarType type;
-  const c10::Device device;
-
-  // See torch/csrc/utils/hash.h for dispatch code.
-  static size_t hash(const BucketKey& key) {
-    return torch::get_hash(key.type, key.device);
-  }
-};
-
-inline bool operator==(const BucketKey& lhs, const BucketKey& rhs) {
-  return lhs.type == rhs.type && lhs.device == rhs.device;
-}
-
 } // namespace
 
 std::vector<std::vector<at::Tensor>> bucketTensors(
@@ -254,84 +232,6 @@ void syncReduction(
   // (NB: original_stream is the current stream PRIOR to the guard.  Might
   // live on a completely different device than our current device here!)
   event.block(cudaGuard.original_stream());
-}
-
-// This is equivalent to take_tensors but returns indices into the
-// tensor list argument for bucket assignment. Also, it is aware
-// of device placement and will not allow buckets to span devices.
-std::vector<std::vector<size_t>> compute_bucket_assignment_by_size(
-    const std::vector<at::Tensor>& tensors,
-    std::vector<size_t> bucket_size_limits) {
-  std::vector<std::vector<size_t>> result;
-  result.reserve(tensors.size());
-
-  // Keep iterator into the size_limit vector by tensor type and device.
-  // This is done so that we can use the consecutive bucket limits per type.
-  std::unordered_map<
-      BucketKey,
-      std::vector<size_t>::iterator,
-      torch::hash<BucketKey>>
-      bucket_size_limit_iterators;
-
-  // Local accumulator type for a single bucket.
-  struct BucketAccumulator {
-    std::vector<size_t> indices;
-    size_t size = 0;
-  };
-
-  // Keep vector of indices and size accumulator by tensor type and device.
-  std::unordered_map<BucketKey, BucketAccumulator, torch::hash<BucketKey>>
-      buckets;
-
-  for (size_t i = 0; i < tensors.size(); i++) {
-    const auto& tensor = tensors[i];
-    AT_ASSERTM(!tensor.is_sparse(), "No support for sparse tensors.");
-    auto key = BucketKey(tensor.scalar_type(), tensor.device());
-    auto& bucket = buckets[key];
-    bucket.indices.push_back(i);
-    bucket.size += tensor.numel() * tensor.element_size();
-
-    // Initialize bucket size limit iterator if necessary.
-    if (bucket_size_limit_iterators.count(key) == 0) {
-      bucket_size_limit_iterators[key] = bucket_size_limits.begin();
-    }
-
-    auto& bucket_size_limit_iterator = bucket_size_limit_iterators[key];
-    const auto bucket_size_limit = *bucket_size_limit_iterator;
-    if (bucket.size >= bucket_size_limit) {
-      result.emplace_back(std::move(bucket.indices));
-      bucket = BucketAccumulator();
-
-      // Advance to the next bucket size limit for this type/device.
-      auto next = bucket_size_limit_iterator + 1;
-      if (next != bucket_size_limits.end()) {
-        bucket_size_limit_iterator = next;
-      }
-    }
-  }
-
-  // Add remaining buckets.
-  for (auto& it : buckets) {
-    auto& bucket = it.second;
-    if (!bucket.indices.empty()) {
-      result.emplace_back(std::move(bucket.indices));
-    }
-  }
-
-  // Sort resulting buckets by the minimum tensor index they include.
-  // We assume that the order of the tensors is the order in which they are
-  // used (or the reverse order in which their gradients are produced).
-  // This sorting step ensures that the buckets are ready in consecutive order.
-  std::sort(
-      result.begin(),
-      result.end(),
-      [](const std::vector<size_t>& a, const std::vector<size_t>& b) {
-        const auto amin = std::min_element(a.begin(), a.end());
-        const auto bmin = std::min_element(b.begin(), b.end());
-        return *amin < *bmin;
-      });
-
-  return result;
 }
 
 } // namespace c10d
