+++ torch/csrc/distributed/c10d/ddp.h
@@ -41,8 +41,4 @@ void syncReduction(
     std::vector<at::Tensor>& gradsBatch,
     at::Tensor& gradsBatchCoalesced);
 
-std::vector<std::vector<size_t>> compute_bucket_assignment_by_size(
-    const std::vector<at::Tensor>& tensors,
-    std::vector<size_t> bucket_size);
-
 } // namespace c10d
